Replace realloc() with malloc()
Thought this would be worth trying, in case the conversion of
realloc() to a mere noarch.Malloc() (with no corresponding
noarch.Free()) was somehow causing a problem.

But there's no evidence of any difference in perf analysis, including
VMpeak consumption (and the tests still fail).

(Probably the Go garbage collection was freeing the abandoned storage
effectively enough to avoid any memory issues.)